modified:   README.md 	modified:   p003/Makefile 	modified:   p003/common.h 	modified:   p003/main.c 	deleted:    p003/timer.c 	deleted:    p003/timer.h 	modified:   p003/uart.c 	modified:   p003/uart.h 	new file:   p004/Makefile 	new file:   p004/common.h 	new file:   p004/key.c 	new file:   p004/key.h 	new file:   p004/main.c 	new file:   p004/stc89c5xrc.h 	new file:   p004/uart.c 	new file:   p004/uart.h

+++ p004/uart.c
@@ -19,7 +19,7 @@ void uart_isr(void) __interrupt (4) __using (1)
 	}
 }
 
-void uart_init(unsigned long baudrate)
+void uart_init(u32 baudrate)
 {
 #if ((PARITY_BIT == PARITY_ODD) || \
      (PARITY_BIT == PARITY_EVEN) || \
@@ -64,5 +64,6 @@ void putchar(char c)
 	is_tx_busy = 1;
 	SBUF = ACC;
 }
+
