+++ p004/uart.h
@@ -11,7 +11,7 @@
 #define PARITY_BIT	PARITY_NONE
 
 void uart_isr(void) __interrupt (4) __using (1);
-void uart_init(unsigned long baudrate);
+void uart_init(u32 baudrate);
 
 #endif
